+++ packages/plugin-ext/src/common/plugin-api-rpc.ts
@@ -114,6 +114,7 @@ export interface Plugin {
     model: PluginModel;
     rawModel: PluginPackage;
     lifecycle: PluginLifecycle;
+    isUnderDevelopment: boolean;
 }
 
 export interface ConfigStorage {
@@ -197,7 +198,8 @@ export const emptyPlugin: Plugin = {
             version: 'empty'
         },
         packagePath: 'empty'
-    }
+    },
+    isUnderDevelopment: false
 };
 
 export interface PluginManagerInitializeParams {