+++ packages/plugin-ext/src/common/plugin-protocol.ts
@@ -795,6 +795,7 @@ export interface PluginMetadata {
     host: string;
     model: PluginModel;
     lifecycle: PluginLifecycle;
+    isUnderDevelopment?: boolean;
 }
 
 export const MetadataProcessor = Symbol('MetadataProcessor');